requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioitertools==0.12.0
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.3.0
botocore==1.38.27
certifi==2025.7.9
cftime==1.6.4.post1
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dask==2025.5.1
debugpy==1.8.14
decorator==5.2.1
distributed==2025.5.1
executing==2.2.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
h5netcdf==1.6.3
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
nest-asyncio==1.6.0
netCDF4==1.7.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
rich==14.0.0
s3fs==2025.5.1
scipy==1.16.0
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.1.0
toolz==1.0.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
wrapt==1.17.2
xarray==2025.7.0
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
